Make doxygen work even if it's initialized before boostbook.
The doxygen module is not really standalone, it requires Boost.Book XSL
stylesheets, and so needs to get their location from the boostbook module.
Previously, the doxygen.init module would get the stylesheet dir, and if
it's empty, would not declare some generators. So, if doxygen is initialized
before bookstbook, some conversion won't work.

Now, generators are defined in all cases, but when the action is run, we
check that boostbook was initialized.