Attribute copyright to Douglas Gregor.
The copyright in these files was attributed to William Kempf, but the revision
history shows that he has never touched them and the original file that they
were based on (threads.html) was created by Douglas Gregor.

The copyright was actually added by Hartmut Kaiser, who was adding the
copyright to all the Boost.Thread files and included this one.