Simplify the tests a little:
Add a parameter to random_values to control what sort of values it generates.
This means that instead of using equivalent_object to test collisions (which
was a total hack) we now just need another parameter.

This requires some meta programming to act differently for maps and sets.
Because of this pairs no longer need to be generated so remove the code for
doing that (which doesn't work on some compilers).

Remove the generator object, just call generate directly.

Remove some of the tests using int containers, they didn't really add to
anthing other than the compile time (some tests are timing out).