Use the same css files for much of the documentation. Refs #4339.
Will need updating whenever the css is updated. I'm going to try to
automate that soonish. I was tempted to serve them from the latest
zipfile, but this is a tad bit faster. Another possibility is rewriting
the html to always use the same css files, this could be helpful since
they'd only be downloaded once but right now I don't want add any more
complexity there.

This is only for css files in doc/src/*.css and doc/html/*.css.