Remove spirit 2 quickbook branch.
I haven't been keeping it up to date. Might port to spirit 2 again in the
future, it should be a lot easier since I'm more experienced with spirit 2 now,
and I've incorporated a lot of the changes into the main version of quickbook
so it's more like a spirit 2 parser nowadays anyway.